pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geminiasr"
version = "1.2.0"
description = "Transcribe video/audio to SRT using Google Gemini API"
readme = "README.md"
authors = [{ name = "cxyfer", email = "xyferim@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["gemini", "asr", "srt", "transcription", "subtitle"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-genai>=1.13.0",
    "moviepy==1.0.3",
    "python-dotenv>=1.1.0",
    "tomli>=1.2.0",
]

[project.urls]
Repository = "https://github.com/cxyfer/GeminiASR"
Issues = "https://github.com/cxyfer/GeminiASR/issues"
Changelog = "https://github.com/cxyfer/GeminiASR/blob/main/CHANGELOG.md"

[project.scripts]
geminiasr = "geminiasr.cli:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/geminiasr"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=geminiasr --cov-report=term-missing --cov-report=html"
python_files = ["test_*.py", "*_test.py"]
